+++ src/frozen/lockable.py
@@ -53,15 +53,24 @@ class LockableClass(ClassDecorator):
 		"""
 		if lock_permissions is None:
 			raise ValueError(Errors.NoLocksDefined.format(self._decorator_function))
+		elif isinstance(lock_permissions, dict):
+			lock_permissions = lock_permissions.items()
+
+		lock_permissions = dict((k, set(v)) for k, v in lock_permissions)
 
 		if unlock_permissions is None:
 			unlock_permissions = lock_permissions
+		else:
+			if isinstance(unlock_permissions, dict):
+				unlock_permissions = unlock_permissions.items()
+
+			unlock_permissions = dict((k, set(v)) for k, v in unlock_permissions)
 
 		self._locks = set(lock_permissions.keys() | unlock_permissions.keys())
 		self._lock_permissions: typing.Dict = defaultdict(lambda: None, lock_permissions)
 		self._unlock_permissions: typing.Dict = defaultdict(lambda: None, unlock_permissions)
 
-	def __call__(self, cls):
+	def __call__(self, cls, *_):
 		"""
 		Sanitize the class cls and returns a wrapper
 		:param cls:
@@ -69,11 +78,29 @@ class LockableClass(ClassDecorator):
 		"""
 		super().__call__(cls)
 
+		# TODO: revisit this part
 		for func in inspect.getmembers(cls, lambda x: inspect.isfunction(x)):
 			try:
 				self._locks.update(func.__locks__)
 			except (AttributeError, TypeError):
 				pass
+
+		class LockableView(ClassDecorator.ObjectView):
+			__locks__: set
+
+			def __init__(self, obj):
+				self.__locks__ = obj.__locks__
+				ClassDecorator.ObjectView.__init__(self, obj)
+
+			def lock(self, key: str):
+				raise PermissionError(
+					Errors.ViewMethodNotCallable.format(LockableView.lock.__name__, cls.__qualname__)
+				)
+
+			def unlock(self, key: str):
+				raise PermissionError(
+					Errors.ViewMethodNotCallable.format(LockableView.unlock.__name__, cls.__qualname__)
+				)
 
 		class LockableWrapper(cls, ClassDecorator.ClassWrapper):
 			def __init__(self, *args, **kwargs):
@@ -144,16 +171,22 @@ class LockableClass(ClassDecorator):
 					raise KeyError(Errors.UnrecognizedKey.format(key)) from None
 
 			# noinspection PyMethodParameters
-			def lock(cls_self, key: str):
+			def lock(myself, key: str) -> None:
+				"""
+				Tries to lock the object with `key`;
+				If the locker is not permitted to use the key, raises a `LockError`.
+				:param key: The key to lock the object
+				:return: None
+				"""
 				if key in self._locks:
 					if self._lock_permissions[key] is None:
-						cls_self.__locks__.add(key)
+						myself.__locks__.add(key)
 					else:
 						calling_cls = None
 						found = False
 
 						for _, _class in trace_execution(self._lock_permissions[key]):
-							if _class is not None and not issubclass(_class, ClassDecorator.ClassWrapper):
+							if _class is not None:
 								found = next(
 									(True for c in self._lock_permissions[key] if issubclass(_class, c)),
 									False
@@ -163,25 +196,25 @@ class LockableClass(ClassDecorator):
 									calling_cls = _class
 
 								if found:
-									cls_self.__locks__.add(key)
+									myself.__locks__.add(key)
 									break
 
 						if not found:
-							cls_self.__lock_error__(key, calling_cls)
+							myself.__lock_error__(key, calling_cls)
 				else:
-					cls_self.__lock_key_error__(key)
+					myself.__lock_key_error__(key)
 
 			# noinspection PyMethodParameters
-			def unlock(cls_self, key: str):
+			def unlock(myself, key: str) -> None:
 				if key in self._locks:
 					if self._unlock_permissions[key] is None:
-						cls_self.__locks__.remove(key)
+						myself.__locks__.remove(key)
 					else:
 						calling_cls = None
 						found = False
 
 						for _, _class in trace_execution(self._unlock_permissions[key]):
-							if _class is not None and not issubclass(_class, ClassDecorator.ClassWrapper):
+							if _class is not None:
 								calling_cls = _class
 								found = next(
 									(True for c in self._unlock_permissions[key] if issubclass(_class, c)),
@@ -189,30 +222,39 @@ class LockableClass(ClassDecorator):
 								)
 
 								if found:
-									cls_self.__locks__.remove(key)
+									myself.__locks__.remove(key)
 									break
 
 						if not found:
-							cls_self.__unlock_error__(key, calling_cls)
+							myself.__unlock_error__(key, calling_cls)
 				else:
-					cls_self.__lock_key_error__(key)
+					myself.__lock_key_error__(key)
 
 			# noinspection PyMethodParameters
-			def locked(cls_self, key: str) -> bool:
+			def locked(myself, key: str) -> bool:
 				"""
 				Returns true if the object is locked with `key`
 				:param key:
 				:return:
 				"""
 				if key in self._locks:
-					return key in cls_self.__locks__
+					return key in myself.__locks__
 				else:
-					cls_self.__lock_key_error__(key)
+					myself.__lock_key_error__(key)
 
 			def view(self):
-				# TODO: Implement this
-				raise NotImplementedError("Not implemented yet. Stay tuned!")
+				if self.__view is None:
+					self.__view = LockableView(self)
 
+				return self.__view
+
+		for lock in self._lock_permissions:
+			self._lock_permissions[lock].add(cls)
+
+		for lock in self._unlock_permissions:
+			self._unlock_permissions[lock].add(cls)
+
+		super().__call__(cls, LockableWrapper)
 		return LockableWrapper
 
 
@@ -231,13 +273,19 @@ class LockableMethod(MethodDecorator):
 	def __call__(self, method: typing.Callable, *_):
 		super().__call__(method)
 
-		def lockable_wrapper(method_self, *args, **kwargs):
-			keys = self._keys.intersection(method_self.__locks__)
+		def lockable_wrapper(myself, *args, **kwargs):
+			if isinstance(myself, ClassDecorator.ObjectView):
+				# noinspection PyProtectedMember
+				locks = myself.__locks__ | myself._ObjectView__obj.__locks__
+			else:
+				locks = myself.__locks__
+
+			keys = self._keys.intersection(locks)
 
 			if keys:
-				method_self.__locked_error__(next(iter(keys)), method)
+				myself.__locked_error__(next(iter(keys)), method)
 			else:
-				return method(method_self, *args, **kwargs)
+				return method(myself, *args, **kwargs)
 
 		return super().__call__(method, lockable_wrapper)